feat(github): Adding membership.yml manifest
Current tracking of reviewers & approvers is spread in a bunch of places.
The desire is to move this into automation sourced from a file in the community repo.
This change does not include any automation, just provides an initial step.

The format on this file may change in the future, it's not prescribing anything.
Upon merging, the `reviewers.md` and `approvers.md` files in `spinnaker/spinnaker` will be deleted.

This also is another step towards cleanup of GitHub groups after splitting from Netflix ownership.