feat(rfc): RFC to replace Halyard
The current recommended way to deploy Spinnaker is by using Halyard.
While Halyard has greatly simplified some aspects of installing and
configuring Spinnaker, it has a number of shortcomings that are
particularly apparent when deploying to Kubernetes. This document
proposes replacing Halyard with a smaller tool that integrates better
into the Kubernetes ecosystem.