Add links to a time zone converter.
Someone in New Zealand was confused about what "Pacific" and "Eastern"
meant. I can't put a UTC time because the UTC time changes whenever we
shift to/from DST.

So I settled on US/Pacific, which should be unambiguous (as opposed to
"PST", which is both confusing and also technically only refers to the
winter time zone) and added a link to a time zone converter.