feat(rfc): Updates to kleat RFC (#144)
* style(rfc): Wrap text

This commit just re-wraps the text; there is no change in content.

* feat(rfc): Updates to kleat RFC

This commit updates the kleat RFC with progress since it was
initially added. In particular:
* Updates the timeline from our (in retrospect) wildly optimistic
estimate. Things are still on track for that plan, just shifted
back a couple of releases (meaning initial implementation next
release).
* There are a few places where we think that minor breaking changes
to the Hal config upon upgrading to kleat are warranted. Updates the
RFC to specify these (and links out to a list of the current ones).
* The initial RFC was vague about how to handle files referenced
by the config. Updates the RFC to outline how this is being handled
initially (though we definitely want feedback on this from early
adopters).
* Provides a bit more detail on what the kustomize install pathway
looks like (and links out to docs that will be updated more frequently
than this RFC)
* Notes that we are not actively working on a Helm pathway, focusing
instead on kustomize.
* Removes the section about having kleat potentially output a BOM.
After further investigation it is much simpler to just have the
containers tagged with the top-level release insead of having this
level of indirection.